* InfectionStatusViewModel.cs: * ILocalNotificationsManager.cs: * LocalNotificationsManager.cs: * spinner_dialog.xml: * iOSLocalNotificationsManager.cs: * HoursPickerModel.cs: * styles_infection_status.xml: * InfectionStatus.storyboard: * InfectionStatusActivity.cs: * InfectionStatusViewController.cs: * pause_exposure_notification_layout.xml: * InfectionStatusViewController.designer.cs: [Issue 179] Functionality   to turn off the Exposure Notifications for a specified time
* WelcomePageTwoViewController.designer.cs:

+++ NDB.Covid19/NDB.Covid19.Droid/Utils/LocalNotificationsManager.cs
@@ -85,7 +85,7 @@ namespace NDB.Covid19.Droid.Utils
             notificationManager?.CreateNotificationChannel(permissionsChannel);
         }
 
-        public void GenerateLocalNotification(NotificationViewModel notificationViewModel, int triggerInSeconds)
+        public void GenerateLocalNotification(NotificationViewModel notificationViewModel, long triggerInSeconds)
         {
             BroadcastNotification(notificationViewModel, NotificationType.Local);
         }
